Extended The User Model in 'core' app
