+++ part-ii/current/storefront2/core/models.py
@@ -1,0 +1,6 @@
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+# Create your models here.
+class User(AbstractUser):
+    email = models.EmailField(unique=True)
