+++ part-ii/current/storefront2/core/views.py
