--- part-ii/current/storefront2/store_custom/admin.py
@@ -1,18 +1,0 @@
-from store.models import Product
-from django.contrib import admin
-from django.contrib.contenttypes.admin import GenericTabularInline
-from store.admin import ProductAdmin
-from tags.models import TaggedItem
-
-
-class TagInline(GenericTabularInline):
-    autocomplete_fields = ['tag']
-    model = TaggedItem
-
-
-class CustomProductAdmin(ProductAdmin):
-    inlines = [TagInline]
-
-
-admin.site.unregister(Product)
-admin.site.register(Product, CustomProductAdmin)
