+++ part-ii/current/storefront2/storefront/settings.py
@@ -42,9 +42,9 @@ INSTALLED_APPS = [
     'playground',
     'debug_toolbar',
     'store',
-    'store_custom',
     'tags',
-    'likes'
+    'likes',
+    'core',
 ]
 
 MIDDLEWARE = [
@@ -147,3 +147,7 @@ REST_FRAMEWORK = {
     # 'PAGE_SIZE': 10,
     # 'DEFAULT_PAGINATION_CLASS': 'rest_framework.pagination.PageNumberPagination',
 }
+
+#* Extended the User model from in 'core.models'
+#* That's why referencing it here.
+AUTH_USER_MODEL = 'core.User'
